Minor code refactor + fixed more import changes
